// src/front/src/api/identity/models/teamResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { TeamType } from "./teamType";

export interface TeamResponse {
	id?: number;
	/** @nullable */
	gid?: string | null;
	structureId?: number;
	/** @nullable */
	sbuName?: string | null;
	/** @nullable */
	code?: string | null;
	/** @nullable */
	name?: string | null;
	type?: TeamType;
	isActive?: boolean;
	createdAt?: string;
	/** @nullable */
	modifiedAt?: string | null;
}